requirements.txt
aiofiles == 22.1.0
aiohttp == 3.8.3
aiosignal == 1.3.1
aiostream == 0.4.5
async-timeout == 4.0.2
attrs == 22.1.0
beautifulsoup4 == 4.11.1
certifi == 2020.12.5
chardet == 3.0.4
charset-normalizer == 2.1.1
click == 8.1.3
dill == 0.3.3
ds4biz-format-parsers == 0.3.0
frozenlist == 1.3.3
html2text == 2020.1.16
httptools == 0.5.0
idna == 2.10
joblib == 1.2.0
langdetect == 1.0.9
Levenshtein == 0.20.8
loko-extensions == 0.0.2
lxml == 4.9.1
multidict == 6.0.2
nltk == 3.7
numpy == 1.23.4
opencv-python == 4.6.0.66
packaging == 21.3
Pillow == 9.3.0
PyMuPDF == 1.21.0
pyparsing == 3.0.9
PyPDF2 == 2.11.1
pytesseract == 0.3.10
python-docx == 0.8.11
python-Levenshtein == 0.20.8
PyYAML == 6.0
rapidfuzz == 2.13.2
rarfile == 4.0
regex == 2022.10.31
requests == 2.25.0
sanic == 22.9.1
sanic-openapi == 21.12.0
sanic-routing == 22.8.0
six == 1.16.0
soupsieve == 2.3.2.post1
tqdm == 4.64.1
ujson == 5.5.0
urllib3 == 1.26.2
uvloop == 0.17.0
websockets == 10.4
XlsxWriter == 3.0.3
yarl == 1.8.1
